Implement new SignedExtension logic to validate unsigned transactions
With the new rules SignedExtension::validate_unsigned will return an error
by default, specifically UnknownTransaction::NoUnsignedValidator to indicate
that unsigned validation was not implemented.

Notice: in this commit, the error-by-default rule is not enabled
so as to keep the tests from failing. It is covered by new tests.

For SignedExtension tuples, there is a new logic as well:

- If all return NoUnsignedValidator (none of them did the check),
  it is considered invalid and Err(NoUnsignedValidator) is returned.

- If one returns an Err(_), it is considered invalid and that error is
  returned. The first error occurred will be returned.

- If at least one returns Ok(_) and no error is returned, it is
  considered valid.

This behaviour applies to validate_unsigned and pre_dispatch_unsigned.
The difference is that pre_dispatch_unsigned will create a Pre::default()
for each NoUnsignedValidator returned (assuming the call succeeds).